fix weird chromium behavior
On Chromium, the paper is unselectable,
and at many similar angles it is partially unselectable.
It can be scrolled in its padding,
and if it's changed to overflow text, it's selectable.

This is fixed by having an inner element be the scroll container.
Also, some styles are updated for consistency.